x64 Phase 2b (1/N): cedtDoc.h + cedtElement.h + cedtDocAnal.cpp
First chunk of the 737-warning size_t-narrowing cleanup. Knocked 519
warnings off the Debug-KR|x64 count (756 → 237) by tackling the
three highest-multiplier sites.

cedtDoc.h, cedtElement.h: MFC's CList::GetCount() returns INT_PTR
on x64. Four inline accessors in cedtDoc.h (GetLineCount,
GetLastIdxY, GetUndoBufferCount, GetRedoBufferCount) and one in
cedtElement.h (CMacroBuffer::GetBufferCount) declared themselves as
INT, so every TU that included these headers re-emitted the
INT_PTR→INT narrowing warning at every use of the accessor.
410 warnings disappeared by adding an explicit (INT) cast inside
the five inlines — the public contract stays INT (line/word counts
fit comfortably) and the narrowing is now stated at the boundary
instead of implicitly.

cedtDocAnal.cpp: 113 warnings, almost all from one repeated pattern.
The file-local helpers _WordFound and _WordFoundExtended took
SHORT siIndex / SHORT siLength parameters and were called ~50 times
with pointer-arithmetic expressions like (beg-str) and (fwd-beg).
Pointer subtraction is INT_PTR on x64, which then narrows to SHORT
at every call site — same explicit-narrowing-at-the-boundary
problem as the inlines above. Bumped the two helpers'
siIndex/siLength parameters from SHORT to INT_PTR and moved the
SHORT cast inside the helper body (where it's safe — the analyzer
caps line length at MAX_STRING_LENGTH = 32767, well within SHORT's
range). Added a comment so the next reader knows the cap is what
makes the narrowing safe.

Same file: the strlen() calls in InitAnalyze (lenHEX / lenHRD /
lenLF1 ... lenR2E) now wrap the strlen result in (INT) since these
are short fixed strings (language-spec tokens) where the int
fits. Single bKEY assignment from CMap::GetCount() also gets a
(BOOL) cast.

After this commit Debug-KR|x64 clean rebuild is at:
  errors 0, warnings 237 (down from 756 last session).

Top remaining files: Utility.cpp (32), cedtElement.cpp (22),
RegExp.cpp (14), PathName.cpp (14), FileWndDirectory.cpp (11),
cedtViewCommand.cpp (10), cedtAppConf.cpp (10), evaluate.cpp (10).
Then ~30 files with under 10 each. The rest of Phase 2b will work
through them in chunks of roughly that order.

cedt_kr.exe x64 Release still boots — quick smoke check after this
commit passed (title "크림슨에디터 - [cedtapp.cpp]" came up in the
earlier verify; nothing in this commit could regress that).

+++ src/core/cedtElement.h
@@ -485,7 +485,7 @@ public:
 	CMacroBuffer();
 	virtual ~CMacroBuffer() {}
 
-	INT GetBufferCount() { return m_lstAction.GetCount(); }
+	INT GetBufferCount() { return (INT)m_lstAction.GetCount(); }
 	CString GetHotKeyText();
 
 	BOOL StreamSave(ofstream & fout);
